Correct the hour comparison for the alarm, hour chime and night light features.
The CurrentHour takes values of between 1-12 when running in 12Hr mode.
The alarm, hour chime and night light hours are stored in FlashConfig in 24 hr mode with values of 0-23. Use the 24hr values in CurrentHourSetting for the comparisons.
This corrects the failures of afternoon alarmsand double triggering when running in 12Hr mode.